examples/bitcoin-price-api-imperative-part-a/requirements-gcp.txt
# GCP Cloud Function Requirements
# Core dependencies
requests>=2.32.5
pandas>=2.0.0

# Google Cloud
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
google-cloud-logging>=3.5.0

# Optional: Enhanced error handling
tenacity>=8.2.0

# Optional: Data validation
pydantic>=2.0.0
